Update to fix a bug for interfaces
The bug was caused by trying to execute an event and specifying the
type. This meant that you could not pass in the generic parameter of the
event you wanted to execute. I changed it to get the type from the
generic parameter that was passed in and this seems to have fixed the
issue