Updates on reflection and factories
Removed a bunch of the reflection and type creation/caching as .net
caches generic type creation. Added an IFactory and Factory that is used
to wrap the single and multi instance factories and take care of all the
casting as needed.

The factory has try methods, i havnt used them yet but it may be an idea
to use these to check if an instance is available just in case the
implementation of SIF or MIF delegates throw and exception or don't
return a given type.